Add ID checking to UsersController with Tests

+++ Api/Controllers/UsersController.cs
@@ -43,6 +43,7 @@ namespace SourceName.Api.Controllers
         public IActionResult Authenticate([FromBody] AuthenticateUserRequest request)
         {
             var token = _userAuthenticationService.Authenticate(request.Username, request.Password);
+
             if (string.IsNullOrWhiteSpace(token))
             {
                 _logger.LogInformation($"Invalid user login. Username: {request.Username}");
@@ -50,6 +51,7 @@ namespace SourceName.Api.Controllers
             }
 
             var user = _userService.GetByUsername(request.Username);
+
             return Ok(new AuthenticateUserResponse
             {
                 FirstName = user.FirstName,
@@ -62,14 +64,36 @@ namespace SourceName.Api.Controllers
         [HttpPost("register")]
         public IActionResult Register([FromBody] CreateUserRequest request)
         {
+            if (_userService.GetByUsername(request.Username) == null)
+            {
+                var errorResponse = new CreateUserResponse
+                {
+                    IsUserCreated = false,
+                    Message = "Username Already Exists"
+                };
+                return Ok(errorResponse);
+            }
+
             var newUser = _userService.CreateUser(_mapper.Map<User>(request));
             _logger.LogInformation($"New User being created: {newUser.Id}");
-            return CreatedAtAction(nameof(Authenticate), _mapper.Map<UserResource>(newUser));
+
+            var response = new CreateUserResponse
+            {
+                IsUserCreated = true,
+                UserResource = _mapper.Map<UserResource>(newUser)
+            };
+
+            return CreatedAtAction(nameof(Authenticate), response);
         }
 
         [HttpDelete("{id}")]
         public IActionResult DeleteUser([FromRoute] int id)
         {
+            if (_userService.GetById(id) == null)
+            {
+                return NotFound();
+            }
+
             _logger.LogInformation($"User {id} is being deleted.");
             _userService.DeleteUser(id);
             return NoContent();
@@ -85,6 +109,10 @@ namespace SourceName.Api.Controllers
         public IActionResult GetById([FromRoute] int id)
         {
             var user = _userService.GetById(id);
+            if (user == null)
+            {
+                return NotFound();
+            }
             return Ok(_mapper.Map<UserResource>(user));
         }
 
@@ -99,6 +127,11 @@ namespace SourceName.Api.Controllers
         [HttpPut("{id}")]
         public IActionResult UpdateUser([FromRoute] int id, [FromBody] UpdateUserRequest request)
         {
+            if (_userService.GetById(id) == null)
+            {
+                return NotFound();
+            }
+
             _logger.LogInformation($"User is being updated: {id}");
             var user = _mapper.Map<User>(request);
             user.Id = id;
@@ -108,6 +141,11 @@ namespace SourceName.Api.Controllers
         [HttpPatch("{id}/password")]
         public IActionResult UpdatePassword([FromRoute] int id, [FromBody] UpdatePasswordRequest request)
         {
+            if (_userService.GetById(id) == null)
+            {
+                return NotFound();
+            }
+
             _logger.LogInformation($"User {id} is updating their password.");
             return Ok(_mapper.Map<UserResource>(
                 _userService.UpdateUserPassword(id, request.Password)));
